requirements.txt
lightning==2.0.0
omegaconf==2.3.0
jaxtyping
typeguard
git+https://github.com/KAIR-BAIR/nerfacc.git@v0.5.2
# git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers==0.25
transformers==4.28.1
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes
wandb
git+https://github.com/ashawkey/envlight.git
torchmetrics
pytorch-lightning==2.0.0
huggingface_hub==0.24.7
numpy==1.26.4
gradio==2.9.4

# # 3d gaussian
plyfile
faiss-gpu
# plyfile

# clipscore
git+https://github.com/openai/CLIP.git
scikit-learn

# diffmc
diso
einops
